some tests, some refactoring

+++ tests/integration_test.rs
@@ -1,13 +1,14 @@
 extern crate openai_api;
 use std::fs;
+use file_diff::{diff, diff_files};
 use openai_api::{ByUrlRequest, DownloadRequest, FormRequest, GetRequest, JsonRequest, OpenAiClient};
 use openai_api::chat::structs::*;
 use serde::Deserialize;
 use openai_api::completion::structs::{CompletionRequest};
 use openai_api::edit::structs::EditRequest;
 use openai_api::embeddings::structs::EmbeddingRequest;
-use openai_api::files::structs::{FileDeleteRequest, FileDownloadRequest, FileInfoRequest, FilesResponse, FileUploadRequest};
-use openai_api::structs::{Input, ModelsResponse};
+use openai_api::files::structs::{FileDeleteRequest, FileDownloadRequest, FileInfoRequest, FileListResponse, FileUploadRequest};
+use openai_api::structs::{Input, ModelListResponse};
 
 #[derive(Deserialize)]
 struct Config{
@@ -64,7 +65,7 @@ async fn completion()-> Result<(),anyhow::Error> {
 #[tokio::test]
 async fn models()-> Result<(),anyhow::Error> {
     let client = get_client();
-    let response = ModelsResponse::get(&client).await?;
+    let response = ModelListResponse::get(&client).await?;
     dbg!(response);
     Ok(())
 }
@@ -84,14 +85,14 @@ async fn embeddings()-> Result<(),anyhow::Error> {
 
 
 #[tokio::test]
-async fn files() -> Result<(),anyhow::Error> {
+async fn file_add() -> Result<(),anyhow::Error> {
     let client = get_client();
     //upload file
-    let file = FileUploadRequest::with_str("Cargo.toml","fine-tune");
+    let file = FileUploadRequest::with_str("fine-tune.json","fine-tune");
     let response = file.run(&client).await?;
     dbg!(&response);
-    //ist uploaded files
-    let files = FilesResponse::get(&client).await?;
+    //list uploaded files
+    let files = FileListResponse::get(&client).await?;
     dbg!(files);
     //get info about single file
     let info_request=FileInfoRequest{
@@ -99,13 +100,61 @@ async fn files() -> Result<(),anyhow::Error> {
     };
     let info= info_request.run(&client).await?;
     dbg!(&info);
+
     //download file
-    let download_request:FileDownloadRequest = info.clone().into();
+    // IMPORTANT! downloading files are disabled for free accounts, so this wont work on free account
+
+    /*let download_request:FileDownloadRequest = info.clone().into();
     download_request.download_to_file(&client,"fine-tune2.json").await?;
+    if !diff("fine-tune.json","fine-tune2.json"){
+        panic!("downloaded file are not the same as uploaded file")
+    }
+    fs::remove_file("fine-tune2.json")?;
+    */
+
     //delete file
+    //also will not work immediately, because
     let delete_request:FileDeleteRequest = info.clone().into();
     let delete_result = delete_request.run(&client).await?;
     dbg!(delete_result);
-    fs::remove_file("fine-tune2.json")?;
+
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn file_delete() -> Result<(),anyhow::Error> {
+    let client = get_client();
+    //list uploaded files
+    let files = FileListResponse::get(&client).await?;
+    dbg!(&files);
+
+    //delete all uploaded  files
+    // IMPORTANT! will not work immediately after uploading, because openai does some processing on uploaded files
+    for file in files.data{
+        let delete_request:FileDeleteRequest = file.clone().into();
+        let delete_result = delete_request.run(&client).await?;
+        dbg!(delete_result);
+    }
+    Ok(())
+}
+
+
+
+
+#[tokio::test]
+async fn fine_tunes() -> Result<(),anyhow::Error> {
+    let client = get_client();
+    //upload fine-tune file
+    let file = FileUploadRequest::with_str("fine-tune.json","fine-tune");
+    let response = file.run(&client).await?;
+    dbg!(&response);
+
+
+
+    //delete file
+    let delete_request:FileDeleteRequest = response.clone().into();
+    let delete_result = delete_request.run(&client).await?;
+    dbg!(&delete_result);
     Ok(())
 }
